[DATA] Comment Entity 수정

+++ src/main/java/com/hyun/twitter/entity/Comment.java
@@ -4,34 +4,34 @@ import java.time.LocalDateTime;
 
 public class Comment {
 
-    private Integer commentId;
-    private Integer postId;
-    private Integer userId;
+    private Long commentId;
+    private Long postId;
+    private Long userId;
     private String content;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
-    public Integer getCommentId() {
+    public Long getCommentId() {
         return commentId;
     }
 
-    public void setCommentId(Integer commentId) {
+    public void setCommentId(Long commentId) {
         this.commentId = commentId;
     }
 
-    public Integer getPostId() {
+    public Long getPostId() {
         return postId;
     }
 
-    public void setPostId(Integer postId) {
+    public void setPostId(Long postId) {
         this.postId = postId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
